splits out gossip CrdsData out of crds_value.rs
Reorganizing this code in preparation of upcoming changes to CrdsValue
(de)serialization.
