+++ core/src/cluster_slots_service.rs
@@ -182,7 +182,7 @@ impl ClusterSlotsService {
 mod test {
     use {
         super::*,
-        solana_gossip::{cluster_info::Node, crds_value::LowestSlot},
+        solana_gossip::{cluster_info::Node, crds_data::LowestSlot},
         solana_sdk::signature::{Keypair, Signer},
         solana_streamer::socket::SocketAddrSpace,
     };
