+++ gossip/src/lib.rs
@@ -5,6 +5,7 @@ pub mod cluster_info;
 pub mod cluster_info_metrics;
 pub mod contact_info;
 pub mod crds;
+pub mod crds_data;
 pub mod crds_entry;
 pub mod crds_gossip;
 pub mod crds_gossip_error;
