+++ gossip/src/restart_crds_values.rs
@@ -1,5 +1,5 @@
 use {
-    crate::crds_value::{new_rand_timestamp, sanitize_wallclock},
+    crate::crds_data::{new_rand_timestamp, sanitize_wallclock},
     bv::BitVec,
     itertools::Itertools,
     rand::Rng,
@@ -219,7 +219,8 @@ mod test {
         super::*,
         crate::{
             cluster_info::MAX_CRDS_OBJECT_SIZE,
-            crds_value::{CrdsData, CrdsValue, CrdsValueLabel},
+            crds_data::CrdsData,
+            crds_value::{CrdsValue, CrdsValueLabel},
         },
         bincode::serialized_size,
         solana_sdk::{signature::Signer, signer::keypair::Keypair, timing::timestamp},
@@ -366,7 +367,7 @@ mod test {
         };
         assert_eq!(fork.sanitize(), Ok(()));
         assert_eq!(fork.observed_stake, 800_000);
-        fork.wallclock = crate::crds_value::MAX_WALLCLOCK;
+        fork.wallclock = crate::crds_data::MAX_WALLCLOCK;
         assert_eq!(fork.sanitize(), Err(SanitizeError::ValueOutOfBounds));
     }
 }
