+++ gossip/tests/crds_gossip.rs
@@ -8,11 +8,12 @@ use {
         cluster_info_metrics::GossipStats,
         contact_info::ContactInfo,
         crds::GossipRoute,
+        crds_data::CrdsData,
         crds_gossip::*,
         crds_gossip_error::CrdsGossipError,
         crds_gossip_pull::{CrdsTimeouts, ProcessPullStats, CRDS_GOSSIP_PULL_CRDS_TIMEOUT_MS},
         crds_gossip_push::CRDS_GOSSIP_PUSH_MSG_TIMEOUT_MS,
-        crds_value::{CrdsData, CrdsValue, CrdsValueLabel},
+        crds_value::{CrdsValue, CrdsValueLabel},
         ping_pong::PingCache,
     },
     solana_rayon_threadlimit::get_thread_count,
