+++ turbine/src/cluster_nodes.rs
@@ -9,8 +9,9 @@ use {
         cluster_info::ClusterInfo,
         contact_info::{ContactInfo, Protocol},
         crds::GossipRoute,
+        crds_data::CrdsData,
         crds_gossip_pull::CRDS_GOSSIP_PULL_CRDS_TIMEOUT_MS,
-        crds_value::{CrdsData, CrdsValue},
+        crds_value::CrdsValue,
         weighted_shuffle::WeightedShuffle,
     },
     solana_ledger::shred::ShredId,
